+++ frontend/src/App.js
@@ -3,13 +3,13 @@ import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { Switch, Route } from 'react-router-dom';
 
-import { AuthRoute, ProtectedRoute } from './components/Routes';
-import NavBar from './components/NavBar';
+import { AuthRoute, ProtectedRoute } from './components/Routes/Routes';
+import NavBar from './components/NavBar/NavBar';
 
-import MainPage from './components/MainPage';
+import MainPage from './components/MainPage/MainPage';
 import LoginForm from './components/SessionForms/LoginForm';
 import SignupForm from './components/SessionForms/SignupForm';
-import Map from './components/Map';
+import Map from './components/Map/Map';
 
 import { getCurrentUser } from './store/session';
