+++ frontend/src/components/ProfilePage/ProfilePage.js
@@ -1,0 +1,5 @@
+const ProfilePage = () => {
+
+}
+
+export default ProfilePage;
