--- frontend/src/components/ProfilePage/index.js
@@ -1,5 +1,0 @@
-const ProfilePage = () => {
-
-}
-
-export default ProfilePage;
